restructure components into folders
