+++ react-app/src/App.js
@@ -3,13 +3,14 @@ import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import LoginForm from './components/auth/LoginForm';
 import SignUpForm from './components/auth/SignUpForm';
-import NavBar from './components/NavBar';
+import NavBar from './components/NavBar/NavBar';
 import ProtectedRoute from './components/auth/ProtectedRoute';
 import UsersList from './components/UsersList';
 import User from './components/User';
 import { authenticate } from './store/session';
-import HomePage from './components/HomePage';
-import Courses from './components/Courses'
+import HomePage from './components/HomePage/HomePage';
+import Courses from './components/GameCourseListing/Courses'
+import CourseView from './components/CourseView/CourseView'
 
 function App() {
   const [loaded, setLoaded] = useState(false);
@@ -47,6 +48,9 @@ function App() {
         </Route>
         <Route path='/games/:gameId' exact={true}>
           <Courses />
+        </Route>
+        <Route path='/course/:courseId' exact={true}>
+          <CourseView />
         </Route>
       </Switch>
     </BrowserRouter>
