+++ react-app/src/components/NavBar/NavBar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import LogoutButton from './auth/LogoutButton';
+import LogoutButton from '.././auth/LogoutButton';
 import './NavBar.css'
 
 const NavBar = () => {
